+++ src/main.js
@@ -1,14 +1,10 @@
 import './assets/css/Login.css'
-import './assets/css/Password_Setting.css'
-import './assets/css/Find_Password.css'
-import './assets/css/Payment_Method.css'
+import './assets/css/PasswordSetting.css'
+import './assets/css/FindPassword.css'
+import './assets/css/PaymentMethod.css'
 import './assets/css/SignUp.css'
 import './assets/css/Authenticate.css'
-// import './assets/css/HotelSearch.css'
-// import './assets/css/HotelListing.css'
 import './assets/css/HotelAccommodation.css'
-
-
 
 import { createApp } from 'vue'
 import App from './App.vue'
