+++ src/views/PasswordSetting.vue
